feat(A2-458): add delete docs functionality

+++ packages/appeals-service-api/src/routes/v2/appellant-submissions/service.js
@@ -2,6 +2,9 @@ const Repo = require('./repo');
 const { rules } = require('@pins/business-rules');
 const { mapTypeCodeToAppealId } = require('@pins/common');
 const repo = new Repo();
+const { docsApiClient } = require('../../../doc-client/docs-api-client');
+const ApiError = require('#errors/apiError');
+const logger = require('#lib/logger');
 
 /**
  * @typedef {import('@prisma/client').AppellantSubmission} AppellantSubmission
@@ -29,28 +32,45 @@ exports.post = async ({ userId, data }) => {
  * @returns {Promise<string>}
  */
 exports.deleteOldSubmissions = async () => {
-	const nonSubmittedSubmissions = await repo.getNonSubmittedSubmissions();
 	const deletedSubmissions = [];
-
-	for (const submission of nonSubmittedSubmissions) {
-		const deadlineDate = rules.appeal.deadlineDate(
-			submission.applicationDecisionDate,
-			mapTypeCodeToAppealId(submission.appealTypeCode),
-			submission.applicationDecision
-		);
-
-		const currentDate = new Date();
-		const threeMonthsPastDeadline = new Date(deadlineDate);
-		threeMonthsPastDeadline.setMonth(threeMonthsPastDeadline.getMonth() + 3);
-
-		if (currentDate > threeMonthsPastDeadline) {
-			await repo.deleteLinkedRecords(submission.id);
-
-			await repo.deleteSubmission(submission.id);
-
-			deletedSubmissions.push(submission.id);
+	try {
+		const nonSubmittedSubmissions = await repo.getNonSubmittedSubmissions();
+		if (!nonSubmittedSubmissions || nonSubmittedSubmissions.length === 0) {
+			return 'No non-submitted submissions to delete';
 		}
-	}
 
-	return `Deleted submissions: ${deletedSubmissions.join(', ')}`;
+		for (const submission of nonSubmittedSubmissions) {
+			const deadlineDate = rules.appeal.deadlineDate(
+				submission.applicationDecisionDate,
+				mapTypeCodeToAppealId(submission.appealTypeCode),
+				submission.applicationDecision
+			);
+
+			const currentDate = new Date();
+			const threeMonthsPastDeadline = new Date(deadlineDate);
+			threeMonthsPastDeadline.setMonth(threeMonthsPastDeadline.getMonth() + 3);
+
+			if (currentDate > threeMonthsPastDeadline) {
+				const documents = await repo.getSubmissionDocumentUploads(submission.id);
+				for (const document of documents) {
+					await docsApiClient.deleteDocument(submission.id, document.id);
+				}
+
+				await repo.deleteLinkedRecords(submission.id);
+
+				await repo.deleteSubmission(submission.id);
+
+				deletedSubmissions.push(submission.id);
+			}
+		}
+
+		if (deletedSubmissions.length >= 1) {
+			return `Deleted submissions: ${deletedSubmissions.join(', ')}`;
+		} else {
+			return 'No old non-submitted submissions to delete';
+		}
+	} catch (e) {
+		logger.error({ e }, 'Error deleting old submissions');
+		throw ApiError.withMessage(500, 'Error deleting old submissions');
+	}
 };
